Updated existing tests and added test case for `__name__`
