+++ test/module/core/errors/semantic-errors/invalid-global.ts
@@ -13,8 +13,8 @@ describe("InvalidGlobalError", () => {
 				const programCtx: KipperProgramContext = await compiler.getProgramCtx(parseData, defaultConfig);
 
 				// Duplicate identifier
-				programCtx.registerBuiltIns({ identifier: globalName, params: [], returnType: "void" });
-				programCtx.registerBuiltIns({ identifier: globalName, params: [], returnType: "void" });
+				programCtx.registerBuiltInFunctions({ identifier: globalName, params: [], returnType: "void" });
+				programCtx.registerBuiltInFunctions({ identifier: globalName, params: [], returnType: "void" });
 			} catch (e) {
 				assert.equal((<KipperError>e).constructor.name, "InvalidGlobalError", "Expected different error");
 				assert((<KipperError>e).line !== undefined, "Expected existing 'line' meta field");
